+++ core/src/com/kunal/PlayGround/VariablesForPlayArea.java
@@ -35,7 +35,7 @@ public class VariablesForPlayArea {
 
     //linked list for bodies
     public static LinkedList<Body> CutOutBodies;
-    public static LinkedList<Short> Angle_Of_Shape;
+    public static LinkedList<Float> Angle_Of_Shape;
     public static LinkedList<LinkedList<Byte>> CutoutShapeVertices;
 
     public static byte shapeNumberSelected = 15 ;
@@ -47,10 +47,10 @@ public class VariablesForPlayArea {
         cantuseDots = new LinkedList<Byte>();
         CutOutBodies = new LinkedList<Body>();
         CutoutShapeVertices = new LinkedList<LinkedList<Byte>>();
-        Angle_Of_Shape = new LinkedList<Short>();
+        Angle_Of_Shape = new LinkedList<Float>();
         Sh_pos = new LinkedList<Vector2>();
         Sh_pos.add(new Vector2(640/AllVariables.PPM, -1000/AllVariables.PPM));
-        Angle_Of_Shape.add((short) 180);
+        Angle_Of_Shape.add(180f);
 
         //all big squre Points
         BigSqurePoints[0][0] = 430;
